// frontend/src/types/api.types.ts
export type TaskType = {
    _id: string;
    title: string;
    description?: string;
    project?: {
        _id: string;
        emoji: string;
        name: string;
    };
    priority: TaskPriorityEnumType;
    status: TaskStatusEnumType;
    assignedTo: {
        _id: string;
        name: string;
        profilePicture: string | null;
    } | null;
    createdBy?: string;
    dueDate: string;
    taskCode: string;
    createdAt?: string;
    updatedAt?: string;
};